tests/e2e/requirements.txt
requests==2.31.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-timeout==2.2.0
python-dotenv==1.0.0
aiohttp==3.9.1
tenacity==8.2.3 